requirements.txt
python-telegram-bot==12.8
requests==2.25.0